Move `CoursierLockfileEntry` and `CoursierResolvedLockfile` to `common.py`
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
