+++ src/python/pants/backend/scala/compile/scalac_test.py
@@ -21,8 +21,13 @@ from pants.engine.fs import FileDigest
 from pants.engine.target import CoarsenedTargets
 from pants.jvm import jdk_rules, testutil
 from pants.jvm.compile import CompileResult, FallibleClasspathEntry
-from pants.jvm.resolve.common import ArtifactRequirement, Coordinate, Coordinates
-from pants.jvm.resolve.coursier_fetch import CoursierLockfileEntry, CoursierResolvedLockfile
+from pants.jvm.resolve.common import (
+    ArtifactRequirement,
+    Coordinate,
+    Coordinates,
+    CoursierLockfileEntry,
+    CoursierResolvedLockfile,
+)
 from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
 from pants.jvm.resolve.coursier_setup import rules as coursier_setup_rules
 from pants.jvm.resolve.coursier_test_util import TestCoursierWrapper
