+++ src/python/pants/jvm/resolve/lockfile_metadata.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Iterable, cast
+from typing import TYPE_CHECKING, Any, Iterable, cast
 
 from pants.core.util_rules.lockfile_metadata import (
     LockfileMetadata,
@@ -14,8 +14,11 @@ from pants.core.util_rules.lockfile_metadata import (
     _get_metadata,
     lockfile_metadata_registrar,
 )
-from pants.jvm.resolve.common import ArtifactRequirement
 from pants.util.ordered_set import FrozenOrderedSet
+
+if TYPE_CHECKING:
+    from pants.jvm.resolve.common import ArtifactRequirement
+
 
 _jvm_lockfile_metadata = lockfile_metadata_registrar(LockfileScope.JVM)
